screenvault/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9      # required for FastAPI file uploads
anthropic==0.25.0
Pillow==10.3.0               # thumbnail generation
ocrmac==0.2.0                # Apple Vision OCR wrapper (macOS only)
python-dotenv==1.0.1
